Move SingleProductPage to pages dir
